+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -20,7 +20,7 @@ from azure.cli.command_modules.vm._validators import _get_resource_group_from_va
 from azure.cli.core.commands.validators import validate_file_or_dict, DefaultStr, DefaultInt
 from azure.keyvault import KeyVaultId
 
-from azure.cli.core.commands import LongRunningOperation
+from azure.cli.core.commands import LongRunningOperation, DeploymentOutputLongRunningOperation
 from azure.cli.core.commands.arm import parse_resource_id, resource_id, is_valid_resource_id
 from azure.cli.core.commands.client_factory import get_mgmt_service_client, get_data_service_client
 from azure.cli.core.util import CLIError
@@ -36,6 +36,9 @@ from ._actions import (load_images_from_aliases_doc,
 from ._client_factory import _compute_client_factory
 
 logger = azlogging.get_az_logger(__name__)
+
+_MSI_PORT = 50342
+_MSI_EXTENSION_VERSION = '1.0'
 
 CachingTypes, VirtualHardDisk, VirtualMachineScaleSet, VirtualMachineCaptureParameters, \
     VirtualMachineScaleSetExtension, VirtualMachineScaleSetExtensionProfile = get_sdk(
@@ -1538,13 +1541,15 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
               storage_profile=None, os_publisher=None, os_offer=None, os_sku=None, os_version=None,
               storage_account_type=None, vnet_type=None, nsg_type=None, public_ip_type=None, nic_type=None,
               validate=False, custom_data=None, secrets=None, plan_name=None, plan_product=None, plan_publisher=None,
-              license_type=None):
+              license_type=None, assign_identity=False, identity_scope=None,
+              identity_role=DefaultStr('Contributor'), identity_role_id=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.command_modules.vm._template_builder import (ArmTemplateBuilder, build_vm_resource,
                                                                 build_storage_account_resource, build_nic_resource,
                                                                 build_vnet_resource, build_nsg_resource,
-                                                                build_public_ip_resource, StorageProfile)
+                                                                build_public_ip_resource, StorageProfile,
+                                                                build_msi_role_assignment, build_msi_extension)
 
     from azure.cli.core._profile import CLOUD
 
@@ -1661,6 +1666,14 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
             'product': plan_product
         }
 
+    if assign_identity:
+        vm_resource['identity'] = {"type": "systemAssigned"}
+        role_assignment_guid = str(_gen_guid())
+        master_template.add_resource(build_msi_role_assignment(vm_name, vm_id, identity_role_id,
+                                                               role_assignment_guid, identity_scope))
+        master_template.add_resource(build_msi_extension(vm_name, location, role_assignment_guid, _MSI_PORT,
+                                                         os_type.lower() != 'windows', _MSI_EXTENSION_VERSION))
+
     master_template.add_resource(vm_resource)
 
     template = master_template.build()
@@ -1684,7 +1697,10 @@ def create_vm(vm_name, resource_group_name, image=None, size='Standard_DS1_v2', 
     else:
         LongRunningOperation()(client.create_or_update(
             resource_group_name, deployment_name, properties, raw=no_wait))
-    return get_vm_details(resource_group_name, vm_name)
+    vm = get_vm_details(resource_group_name, vm_name)
+    if assign_identity:
+        setattr(vm, 'identity', _create_identity_object(identity_scope, identity_role, _MSI_PORT))
+    return vm
 
 
 # pylint: disable=too-many-locals, too-many-statements
@@ -1711,14 +1727,17 @@ def create_vmss(vmss_name, resource_group_name, image,
                 load_balancer_type=None, app_gateway_type=None, vnet_type=None,
                 public_ip_type=None, storage_profile=None,
                 single_placement_group=None, custom_data=None, secrets=None,
-                plan_name=None, plan_product=None, plan_publisher=None):
+                plan_name=None, plan_product=None, plan_publisher=None,
+                assign_identity=False, identity_scope=None, identity_role=DefaultStr('Contributor'),
+                identity_role_id=None):
     from azure.cli.core.commands.client_factory import get_subscription_id
     from azure.cli.core.util import random_string, hash_string
     from azure.cli.command_modules.vm._template_builder import (ArmTemplateBuilder, StorageProfile, build_vmss_resource,
                                                                 build_vnet_resource, build_public_ip_resource,
                                                                 build_load_balancer_resource,
                                                                 build_vmss_storage_account_pool_resource,
-                                                                build_application_gateway_resource)
+                                                                build_application_gateway_resource,
+                                                                build_msi_role_assignment)
 
     from azure.cli.core._profile import CLOUD
 
@@ -1902,6 +1921,25 @@ def create_vmss(vmss_name, resource_group_name, image,
             'product': plan_product
         }
 
+    if assign_identity:
+        vmss_resource['identity'] = {"type": "systemAssigned"}
+        role_assignment_guid = str(_gen_guid())
+        master_template.add_resource(build_msi_role_assignment(vmss_name, vmss_id, identity_role_id,
+                                                               role_assignment_guid, identity_scope, False))
+        # pylint: disable=line-too-long
+        vmss_resource['properties']['virtualMachineProfile']['extensionProfile'] = vmss_resource['properties']['virtualMachineProfile'].get('extensionProfile') or {}
+        vmss_resource['properties']['virtualMachineProfile']['extensionProfile']["extensions"] = vmss_resource['properties']['virtualMachineProfile']['extensionProfile'].get('extensions') or []
+        vmss_resource['properties']['virtualMachineProfile']['extensionProfile']["extensions"].append({
+            "name": "MSIExtension",
+            "properties": {
+                "publisher": "Microsoft.ManagedIdentity",
+                "type": "ManagedIdentityExtensionFor" + ('Windows' if os_type.lower() == 'windows' else 'Linux'),
+                "typeHandlerVersion": "1.0",
+                "autoUpgradeMinorVersion": True,
+                'settings': {'port': _MSI_PORT}
+            }
+        })
+
     master_template.add_resource(vmss_resource)
     master_template.add_output('VMSS', vmss_name, 'Microsoft.Compute', 'virtualMachineScaleSets',
                                output_type='object')
@@ -1920,7 +1958,11 @@ def create_vmss(vmss_name, resource_group_name, image,
         return client.validate(resource_group_name, deployment_name, properties, raw=no_wait)
 
     # creates the VMSS deployment
-    return client.create_or_update(resource_group_name, deployment_name, properties, raw=no_wait)
+    deployment_result = DeploymentOutputLongRunningOperation()(
+        client.create_or_update(resource_group_name, deployment_name, properties, raw=no_wait))
+    if assign_identity:
+        deployment_result['vmss']['identity'] = _create_identity_object(identity_scope, identity_role, _MSI_PORT)
+    return deployment_result
 
 
 def create_av_set(availability_set_name, resource_group_name,
@@ -2000,3 +2042,100 @@ def get_vm_format_secret(secrets, certificate_store=None):
                  for _, value in list(grouped_secrets.items())]
 
     return formatted
+
+
+def assign_vm_identity(resource_group_name, vm_name, identity_role='Contributor', identity_role_id=None,
+                       identity_scope=None, port=None):
+    from azure.mgmt.authorization import AuthorizationManagementClient
+    from azure.mgmt.authorization.models import RoleAssignmentProperties
+    from msrestazure.azure_exceptions import CloudError
+    VirtualMachineIdentity = get_sdk(ResourceType.MGMT_COMPUTE, 'VirtualMachineIdentity', mod='models')
+    vm = get_vm(resource_group_name, vm_name)
+    if not vm.identity:
+        logger.info('Enable managed identity...')
+        vm.identity = VirtualMachineIdentity(type='systemAssigned')
+        vm = set_vm(vm)
+    else:
+        logger.info('Managed identity is already enabled')
+
+    # create role assignment
+    logger.info("Creating an assignment with a role '%s' on the scope of '%s'", identity_role, identity_scope)
+    assignments_client = get_mgmt_service_client(AuthorizationManagementClient).role_assignments
+    properties = RoleAssignmentProperties(identity_role_id, vm.identity.principal_id)
+    try:
+        assignments_client.create(identity_scope, _gen_guid(), properties)
+    except CloudError as ex:
+        if ex.message != 'The role assignment already exists.':
+            raise
+        else:
+            logger.info('Role assignment already exists')
+    # provision msi extension
+    port = port or _MSI_PORT
+    ext_name = 'ManagedIdentityExtensionFor' + ('Linux' if _is_linux_vm(vm) else 'Windows')
+    logger.info("Provisioning extension: '%s'", ext_name)
+    poller = set_extension(resource_group_name, vm_name,
+                           publisher='Microsoft.ManagedIdentity',
+                           vm_extension_name=ext_name,
+                           version=_MSI_EXTENSION_VERSION,
+                           settings={'port': port})
+    LongRunningOperation()(poller)
+    return _create_identity_object(identity_scope, identity_role, port)
+
+
+def assign_vmss_identity(resource_group_name, vmss_name, identity_role='Contributor', identity_role_id=None,
+                         identity_scope=None, port=None):
+    from azure.mgmt.authorization import AuthorizationManagementClient
+    from azure.mgmt.authorization.models import RoleAssignmentProperties
+    from msrestazure.azure_exceptions import CloudError
+    VirtualMachineIdentity, UpgradeMode = get_sdk(ResourceType.MGMT_COMPUTE, 'VirtualMachineScaleSetIdentity',
+                                                  'UpgradeMode', mod='models')
+    client = _compute_client_factory()
+    vmss = client.virtual_machine_scale_sets.get(resource_group_name, vmss_name)
+    if not vmss.identity:
+        logger.info('Enable managed identity...')
+        vmss.identity = VirtualMachineIdentity(type='systemAssigned')
+        vmss = client.virtual_machine_scale_sets.create_or_update(resource_group_name, vmss_name, vmss)
+    else:
+        logger.info('Managed identity is already enabled')
+
+    # create role assignment
+    logger.info("Creating an assignment with a role '%s' on the scope of '%s'", identity_role, identity_scope)
+    assignments_client = get_mgmt_service_client(AuthorizationManagementClient).role_assignments
+    properties = RoleAssignmentProperties(identity_role_id, vmss.identity.principal_id)
+    try:
+        assignments_client.create(identity_scope, _gen_guid(), properties)
+    except CloudError as ex:
+        if ex.message != 'The role assignment already exists.':
+            raise
+        else:
+            logger.info('Role assignment already exists')
+    # provision msi extension
+    port = port or _MSI_PORT
+    ext_name = 'ManagedIdentityExtensionFor' + ('Linux' if vmss.virtual_machine_profile.os_profile.linux_configuration
+                                                else 'Windows')
+    logger.info("Provisioning extension: '%s'", ext_name)
+    poller = set_vmss_extension(resource_group_name, vmss_name,
+                                publisher='Microsoft.ManagedIdentity',
+                                extension_name=ext_name,
+                                version=_MSI_EXTENSION_VERSION,
+                                settings={'port': port})
+    LongRunningOperation()(poller)
+    if vmss.upgrade_policy.mode == UpgradeMode.manual:
+        logger.warning("With manual upgrade mode, you will need to run 'az vmss update-instances -g %s -n %s "
+                       "--instance-ids *' to propagate the change", resource_group_name, vmss_name)
+    return _create_identity_object(identity_scope, identity_role, port)
+
+
+def _create_identity_object(identity_scope, identity_role, port):
+    return {
+        'scope': identity_scope,
+        'role': str(identity_role),  # could be DefaultStr, so convert to string
+        'token_endpoint': 'http://localhost:{}/oauth2/token'.format(port),
+        'subscription': re.match(r'/subscriptions/(\w{8}-\w{4}-\w{4}-\w{4}-\w{12})', identity_scope).group(1)
+    }
+
+
+# for injecting test seams to produce predicatable role assignment id
+def _gen_guid():
+    import uuid
+    return uuid.uuid4()